pyproject.toml
[tool.poetry]
name = "git-good-base"
version = "0.1.0"
description = "Find a good commit to base your work on"
authors = ["David Bradford <david.bradford@mongodb.com>"]
packages = [
    { include = "goodbase", from = "src"},
]

[tool.poetry.scripts]
git-good-base = "goodbase.goodbase_cli:main"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8"
"evergreen.py" = { git = "https://github.com/dbradf/evergreen.py.git", branch = "relax-dependecies" }
plumbum = "^1.7.0"
Inject = "^4.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = "^21.7b0"
pytest-black = "^0.3"
pytest-cov = "^2.8"
pytest-flake8 = "^1.0"
pytest-mypy = "^0.8"
mypy = "^0.910"
pytest-pydocstyle = "^2.0"
mongomock = "^3.19"
hypothesis = "^5.16"
pytest-structlog = "^0.3"
pre-commit = "^2.6"
pytest-isort = "^2.0"
flake8-bugbear = "^21.4"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100

[tool.pytest.ini_options]
flake8-ignore = "W605 W503 W291 E203 E501 F821"
addopts = "--flake8 --black --mypy --isort --pydocstyle"
testpaths = [
    "src",
    "tests",
]

[tool.mypy]
ignore_missing_imports = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
disallow_untyped_calls = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
]
ignore_errors = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
